requirements.txt
discord.py==1.3.0
requests==2.22.0
